requirements.txt
autopep8==1.6.0
cycler==0.11.0
fonttools==4.33.3
greenlet==1.1.2
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
patsy==0.5.2
Pillow==9.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.36
statsmodels==0.13.2
threadpoolctl==3.1.0
toml==0.10.2
